Add Output files 2021-10-26
I never actually placed an order but I wanted to see the price of it and
see if I remembered how to order things from JLCPCB.

I noticed that a couple of components needs to be rotated and/or moved
in order to be placed correctly.

I should also add part numbers to the resistors and capacitors in order
to have them mounted as well.